payroll: add get total owed salary function
